Move metadata classes
