+++ src/Sulu/Bundle/AdminBundle/Metadata/Form/Form.php
@@ -9,7 +9,7 @@
  * with this source code in the file LICENSE.
  */
 
-namespace Sulu\Bundle\AdminBundle\ResourceMetadata\Form;
+namespace Sulu\Bundle\AdminBundle\Metadata\Form;
 
 use JMS\Serializer\Annotation\SerializedName;
 use Sulu\Bundle\AdminBundle\ResourceMetadata\Schema\Schema;
