+++ src/Sulu/Bundle/AdminBundle/Metadata/Form/FormMetadataProvider.php
@@ -16,14 +16,6 @@ use Sulu\Bundle\AdminBundle\Exception\MetadataNotFoundException;
 use Sulu\Bundle\AdminBundle\FormMetadata\FormMetadata;
 use Sulu\Bundle\AdminBundle\FormMetadata\FormXmlLoader;
 use Sulu\Bundle\AdminBundle\Metadata\MetadataProviderInterface;
-use Sulu\Bundle\AdminBundle\ResourceMetadata\Form\Field;
-use Sulu\Bundle\AdminBundle\ResourceMetadata\Form\FieldType;
-use Sulu\Bundle\AdminBundle\ResourceMetadata\Form\Form;
-use Sulu\Bundle\AdminBundle\ResourceMetadata\Form\Item;
-use Sulu\Bundle\AdminBundle\ResourceMetadata\Form\Option;
-use Sulu\Bundle\AdminBundle\ResourceMetadata\Form\Section;
-use Sulu\Bundle\AdminBundle\ResourceMetadata\Form\Tag;
-use Sulu\Bundle\AdminBundle\ResourceMetadata\Form\TypedForm;
 use Sulu\Bundle\AdminBundle\ResourceMetadata\Schema\Property;
 use Sulu\Bundle\AdminBundle\ResourceMetadata\Schema\Schema;
 use Sulu\Component\Content\Metadata\BlockMetadata;
